+++ tests/tests.c/findalldevstest.c
@@ -19,6 +19,7 @@
 
 #include <pcap.h>
 
+#include "varattrs.h"
 #include "pcap/funcattrs.h"
 
 static int ifprint(pcap_if_t *d);
@@ -94,7 +95,11 @@ getpass(const char *prompt)
 }
 #endif
 
+#ifdef ENABLE_REMOTE
 int main(int argc, char **argv)
+#else
+int main(int argc _U_, char **argv _U_)
+#endif
 {
   pcap_if_t *alldevs;
   pcap_if_t *d;